(TODO - ticket required to get into release notes) remove default_facts.yml
This commit removes the template-driven default facts, most of which were
duplicating facts already available in rspec-puppet-facts, or workarounds
for Puppet Inc. internal issues.

At the same time, this switches to using rspec-puppet-facts'
`add_custom_fact` method to improve integration with that library.

See the list below for detailed arguments for each change:

* `ipaddress` and `macaddress` facts are supplied by `rspec-puppet-facts`.
* is_pe [TO CONFIRM: is not used in open source modules] and should be handled locally in modules requiring it.
* extra_facts should be supplied in `default_module_facts.yml`, instead of jumping through .sync.yml hoops.
* `facterversion` is provided through the facterdb, and is more accurate than the loaded gem (which is still the ruby facter 2)
* `puppetversion` is automatically set to `Puppet.version` by rspec-puppet-facts.


TODO: what happens with tests that do not use RPF? do we still meed to set default_facts for those?